Module google_taskqueue1_beta2::api

Structs

Task
There is no detailed description.

TaskDeleteCall
Delete a task from a TaskQueue.

TaskGetCall
Get a particular task from a TaskQueue.

TaskInsertCall
Insert a new task in a TaskQueue

TaskLeaseCall
Lease 1 or more tasks from a TaskQueue.

TaskListCall
List Tasks in a TaskQueue

TaskMethods
A builder providing access to all methods supported on task resources. It is not used directly, but through the Taskqueue hub.

TaskPatchCall
Update tasks that are leased out of a TaskQueue. This method supports patch semantics.

TaskQueue
There is no detailed description.

TaskQueueAcl
ACLs that are applicable to this TaskQueue object.

TaskQueueStats
Statistics for the TaskQueue object in question.

TaskUpdateCall
Update tasks that are leased out of a TaskQueue.

Taskqueue
Central instance to access all Taskqueue related resource activities

TaskqueueGetCall
Get detailed information about a TaskQueue.

TaskqueueMethods
A builder providing access to all methods supported on taskqueue resources. It is not used directly, but through the Taskqueue hub.

Tasks
There is no detailed description.

Tasks2
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.