Struct google_taskqueue1_beta2::Task

pub struct Task {
    pub kind: Option<String>,
    pub lease_timestamp: Option<String>,
    pub queue_name: Option<String>,
    pub retry_count: Option<i32>,
    pub tag: Option<String>,
    pub payload_base64: Option<String>,
    pub id: Option<String>,
    pub enqueue_timestamp: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

lease tasks (none)

insert tasks (request|response)

delete tasks (none)

patch tasks (request|response)

list tasks (none)

get tasks (response)

update tasks (request|response)

Fields

kind

The kind of object returned, in this case set to task.

lease_timestamp

Time (in seconds since the epoch) at which the task lease will expire. This value is 0 if the task isnt currently leased out to a worker.

queue_name

Name of the queue that the task is in.

retry_count

The number of leases applied to this task.

tag

Tag for the task, could be used later to lease tasks grouped by a specific tag.

payload_base64

A bag of bytes which is the task payload. The payload on the JSON side is always Base64 encoded.

id

Name of the task.

enqueue_timestamp

Time (in seconds since the epoch) at which the task was enqueued.