Struct google_taskqueue1_beta2::TaskMethods

pub struct TaskMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on task resources. It is not used directly, but through the Taskqueue hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_taskqueue1_beta2 as taskqueue1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use taskqueue1_beta2::Taskqueue;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Taskqueue::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `lease(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.tasks();

Methods

impl<'a, C, A> TaskMethods<'a, C, A>

fn lease(&self, project: &str, taskqueue: &str, num_tasks: i32, lease_secs: i32) -> TaskLeaseCall<'a, C, A>

Create a builder to help you perform the following task:

Lease 1 or more tasks from a TaskQueue.

Arguments

project - The project under which the queue lies.

taskqueue - The taskqueue to lease a task from.

numTasks - The number of tasks to lease.

leaseSecs - The lease in seconds.

fn insert(&self, request: Task, project: &str, taskqueue: &str) -> TaskInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Insert a new task in a TaskQueue

Arguments

request - No description provided.

project - The project under which the queue lies

taskqueue - The taskqueue to insert the task into

fn delete(&self, project: &str, taskqueue: &str, task: &str) -> TaskDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a task from a TaskQueue.

Arguments

project - The project under which the queue lies.

taskqueue - The taskqueue to delete a task from.

task - The id of the task to delete.

fn patch(&self, request: Task, project: &str, taskqueue: &str, task: &str, new_lease_seconds: i32) -> TaskPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update tasks that are leased out of a TaskQueue. This method supports patch semantics.

Arguments

request - No description provided.

project - The project under which the queue lies.

taskqueue - No description provided.

task - No description provided.

newLeaseSeconds - The new lease in seconds.

fn list(&self, project: &str, taskqueue: &str) -> TaskListCall<'a, C, A>

Create a builder to help you perform the following task:

List Tasks in a TaskQueue

Arguments

project - The project under which the queue lies.

taskqueue - The id of the taskqueue to list tasks from.

fn get(&self, project: &str, taskqueue: &str, task: &str) -> TaskGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a particular task from a TaskQueue.

Arguments

project - The project under which the queue lies.

taskqueue - The taskqueue in which the task belongs.

task - The task to get properties of.

fn update(&self, request: Task, project: &str, taskqueue: &str, task: &str, new_lease_seconds: i32) -> TaskUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update tasks that are leased out of a TaskQueue.

Arguments

request - No description provided.

project - The project under which the queue lies.

taskqueue - No description provided.

task - No description provided.

newLeaseSeconds - The new lease in seconds.